Develop

+++ src/main/java/br/com/ucsal/meta/agil/dto/AvaliacaoDTO.java
@@ -9,7 +9,7 @@ public class AvaliacaoDTO {
 	private Long cdAvaliacao;
 	private String nmAvaliacao;
 	private String flAvaliacao;
-	private Double notaAvaliacao;
+	private Integer notaAvaliacao;
 	private Integer cdAplicacao;
 	private Integer cdTime;
 	
@@ -33,10 +33,10 @@ public class AvaliacaoDTO {
 	public void setFlAvaliacao(String flAvaliacao) {
 		this.flAvaliacao = flAvaliacao;
 	}
-	public Double getNotaAvaliacao() {
+	public Integer getNotaAvaliacao() {
 		return notaAvaliacao;
 	}
-	public void setNotaAvaliacao(Double notaAvaliacao) {
+	public void setNotaAvaliacao(Integer notaAvaliacao) {
 		this.notaAvaliacao = notaAvaliacao;
 	}
 	public Integer getCdAplicacao() {
